+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router } from 'react-router-dom'
 import Header from './components/Header.tsx'
 import Footer from './components/Footer.tsx'
-import PostsManagerPage from './pages/PostsManagerPage.tsx'
+import PostsManagerPage from './pages/posts/ui/PostsManagerPage.tsx'
 
 const App = () => {
   return (
